added src files and examples
first commit in feat-v0 branch, adding all src files and initial two examples for ofxGuiJson
