+++ example_audio/addons.make
@@ -1,0 +1,3 @@
+ofxGui
+ofxGuiJson
+ofxMaxim
