+++ example_oscilloscope/addons.make
@@ -1,0 +1,2 @@
+ofxGui
+ofxGuiJson
