+++ example_oscilloscope/src/main.cpp
@@ -2,8 +2,8 @@
 #include "ofApp.h"
 
 //========================================================================
-int main() {
-	ofSetupOpenGL(1024, 768, OF_WINDOW);			// <-------- setup the GL context
+int main( ){
+	ofSetupOpenGL(1024,768,OF_WINDOW);			// <-------- setup the GL context
 
 	// this kicks off the running of my app
 	// can be OF_WINDOW or OF_FULLSCREEN
